social_media_api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.5
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.1
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.26.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
mysql-connector==2.2.9
mysql-connector-python==9.3.0
mysqlclient==2.2.7
openai==1.97.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.2
pyparsing==3.2.3
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
tenacity==8.5.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
waitress==3.0.2
websockets==15.0.1
whitenoise==6.11.0
